Replace SettingOverrideCheckbox in settings views

+++ src/Presentation/SmartStore.Web/Administration/Validators/Settings/RewardPointsSettingsValidator.cs
@@ -9,10 +9,10 @@ namespace SmartStore.Admin.Validators.Settings
     {
         public RewardPointsSettingsValidator(ILocalizationService localizationService)
         {
-            RuleFor(x => x.PointsForPurchases_Awarded).NotEqual((int)OrderStatus.Pending)
+            RuleFor(x => x.PointsForPurchases_Awarded).NotEqual(OrderStatus.Pending)
 				.WithMessage(localizationService.GetResource("Admin.Configuration.Settings.RewardPoints.PointsForPurchases_Awarded.Pending"));
 
-            RuleFor(x => x.PointsForPurchases_Canceled).NotEqual((int)OrderStatus.Pending)
+            RuleFor(x => x.PointsForPurchases_Canceled).NotEqual(OrderStatus.Pending)
 				.WithMessage(localizationService.GetResource("Admin.Configuration.Settings.RewardPoints.PointsForPurchases_Canceled.Pending"));
         }
     }
